+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Fixed bug were map units from polymer with no label or duplicate labels would get dropped from results.
 
 ## [0.4.6] - 2024-10-23